chore: move to sdk-types
