+++ packages/agents/sdk-wrapper/.gitignore
@@ -1,1 +1,1 @@
-types
+src/sdk-types/
