+++ packages/agents/sdk-wrapper/test/sdkRouter.spec.ts
@@ -8,7 +8,7 @@ import {
   SdkAddLiquidityForRouterParams,
   SdkRemoveRouterLiquidityForParams,
   SdkRemoveRouterLiquidityParams,
-} from "../types";
+} from "../src/sdk-types";
 import { expect, mkAddress } from "@connext/nxtp-utils";
 
 const mockConfig = mock.config();
